chore: Move schema model to /shared
