+++ editor.planx.uk/src/@planx/components/List/schemas/Adverts.ts
@@ -1,4 +1,4 @@
-import { Schema } from "@planx/components/List/model";
+import { Schema } from "@planx/components/shared/Schema/model";
 
 export const ProposedAdvertisements: Schema = {
   type: "Proposed advertisements",
