+++ editor.planx.uk/src/@planx/components/List/schemas/GLA/BuildingDetails.ts
@@ -1,4 +1,4 @@
-import { Schema } from "@planx/components/List/model";
+import { Schema } from "@planx/components/shared/Schema/model";
 import { TextInputType } from "@planx/components/TextInput/model";
 
 export const BuildingDetailsGLA: Schema = {
