+++ editor.planx.uk/src/@planx/components/List/schemas/ResidentialUnits/Existing.ts
@@ -1,4 +1,4 @@
-import { Schema } from "../../model";
+import { Schema } from "@planx/components/shared/Schema/model";
 
 export const ResidentialUnitsExisting: Schema = {
   type: "Existing residential unit type",
