+++ editor.planx.uk/src/@planx/components/List/schemas/mocks/GenericUnits.ts
@@ -1,4 +1,4 @@
-import { Schema } from "@planx/components/List/model";
+import { Schema } from "@planx/components/shared/Schema/model";
 
 import { Props } from "../../Public";
